+++ packages/terra-markdown/tests/nightwatch/MarkdownTests.jsx
@@ -1,0 +1,14 @@
+/* eslint-disable import/no-extraneous-dependencies */
+
+import React from 'react';
+import { Link } from 'react-router';
+
+const MarkdownTests = () => (
+  <div>
+    <ul>
+      <li><Link to="/tests/markdown/default">Mock Markdown</Link></li>
+    </ul>
+  </div>
+);
+
+export default MarkdownTests;