+++ packages/terra-site/src/examples/profile-image/ProfileImageDefault.jsx
@@ -6,6 +6,7 @@ const ProfileImageDefault = () => (
   <div>
     <h2>Successful Profile Image</h2>
     <ProfileImage src={exampleProfileImage} width="75" height="75" />
+    <br />
     <h2>Failed Profile Image</h2>
     <ProfileImage src="invalid.jpg" width="75" height="75" />
   </div>