+++ packages/terra-date-picker/CHANGELOG.md
@@ -33,6 +33,9 @@ Unreleased
 ### Fixed
 * Check null date when date entry is cleared
 
+### Added
+* Internationalized this component
+
 1.0.0 - (June 28, 2017)
 ------------------
 Initial stable release