+++ packages/terra-date-picker/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+1.5.0 - (August 8, 2017)
+------------------
+### Changed
+* Change data-class to be name spaced
+
 1.4.0 - (August 1, 2017)
 ------------------
 ### Changed
@@ -32,6 +37,9 @@ Unreleased
 
 ### Fixed
 * Check null date when date entry is cleared
+
+### Added
+* Internationalized this component
 
 1.0.0 - (June 28, 2017)
 ------------------